feat: add endpoint to generate conversation reports
separate reporting from experiments to allow post-conversation reports
