+++ report_templates/directMessageResponses.text.hbs
@@ -1,11 +1,11 @@
 Direct Message Agent Responses Report
 ===========================
 
-Experiment: {{experiment.name}}
+Name: {{experiment.name}}
 {{#if experiment.description}}
 Description: {{experiment.description}}
 {{/if}}
-Experiment Time: {{formatDate experiment.executedAt}}
+Time: {{formatDate experiment.executedAt}}
 Base Conversation ID: {{baseConversation}}
 Result Conversation ID: {{resultConversation}}
