+++ report_templates/periodicResponses.text.hbs
@@ -1,11 +1,11 @@
 Periodic Agent Responses Report
 ===========================
 
-Experiment: {{experiment.name}}
+Name: {{experiment.name}}
 {{#if experiment.description}}
 Description: {{experiment.description}}
 {{/if}}
-Experiment Time: {{formatDate experiment.executedAt}}
+Time: {{formatDate experiment.executedAt}}
 Base Conversation ID: {{baseConversation}}
 Result Conversation ID: {{resultConversation}}
 Unique Participants: {{participantCount}}
