+++ src/services/index.ts
@@ -11,4 +11,5 @@ export { default as pollService } from './poll.service/index.js'
 export { default as experimentsService } from './experiment.service.js'
 export { default as adapterService } from './adapter.service.js'
 export { default as channelService } from './channel.service.js'
+export { default as reportService } from './report.service.js'
 // export { default as exportService } from './export.service.js'
